walletController_almost_done

+++ src/main/java/com/example/trainingsapp/wallet/WalletController.java
@@ -6,18 +6,14 @@ import com.example.trainingsapp.wallet.api.WalletService;
 import com.example.trainingsapp.wallet.api.dto.WalletCreateDTO;
 import com.example.trainingsapp.wallet.api.dto.WalletDTO;
 import com.example.trainingsapp.wallet.api.dto.WalletUpdateDTO;
-import com.example.trainingsapp.wallet.model.Wallet;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
 
 import java.security.Principal;
 import java.util.List;
@@ -27,22 +23,19 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/api/wallets")
 public class WalletController {
-
     @Autowired
-    UserRepository userRepository;
-
+    private UserRepository userRepository;
     @Autowired
-    WalletService walletService;
-
-    @Autowired
-    private static final Logger logger = LoggerFactory.getLogger(WalletController.class);
+    private WalletService walletService;
 
     @PostMapping()
-    public ResponseEntity<WalletDTO> createWallet(@Valid @RequestBody WalletCreateDTO createWalletDTO, Principal principal) {
-
+    public ResponseEntity<WalletDTO> createWallet(@Valid @RequestBody WalletCreateDTO createWalletDTO,
+                                                  Principal principal) {
         String email = principal.getName();
         Optional<User> user = userRepository.findByEmail(email);
-
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Long userID = user.get().getId();
 
         WalletDTO walletDTO = walletService.createWallet(createWalletDTO, userID);
@@ -53,10 +46,12 @@ public class WalletController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteWallet(@Min(1) @NotNull @PathVariable Long id, Principal principal) {
+    public ResponseEntity<String> deleteWalletById(@Min(1) @NotNull @PathVariable Long id, Principal principal) {
         String email = principal.getName();
         Optional<User> user = userRepository.findByEmail(email);
-
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Long userId = user.get().getId();
 
         walletService.deleteWallet(id, userId);
@@ -65,14 +60,13 @@ public class WalletController {
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<WalletDTO> updateWallet(@Min(1) @NotNull @PathVariable Long id, @Valid @RequestBody WalletUpdateDTO walletUpdateDTO,
-                                       Principal principal) {
-
-        System.out.println("Received request to update wallet with id: " + id);
-
+    public ResponseEntity<WalletDTO> updateWalletName(@Min(1) @NotNull @PathVariable Long id,
+                                                      @Valid @RequestBody WalletUpdateDTO walletUpdateDTO, Principal principal) {
         String email = principal.getName();
         Optional<User> user = userRepository.findByEmail(email);
-
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Long userId = user.get().getId();
 
         WalletDTO walletDTO = walletService.updateWallet(id, walletUpdateDTO, userId);
@@ -81,14 +75,16 @@ public class WalletController {
 
     }
 
-    @GetMapping("/")
-    public ResponseEntity<List<Wallet>> getWallets(Principal principal) {
+    @GetMapping()
+    public ResponseEntity<List<WalletDTO>> getWallets(Principal principal) {
         String email = principal.getName();
         Optional<User> user = userRepository.findByEmail(email);
-
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Long userId = user.get().getId();
 
-        List<Wallet> wallets = walletService.getWallets(userId);
+        List<WalletDTO> wallets = walletService.getWallets(userId);
 
         return new ResponseEntity<>(wallets, HttpStatus.OK);
 
@@ -98,7 +94,9 @@ public class WalletController {
     public ResponseEntity<WalletDTO> getWalletById(@Min(1) @NotNull @PathVariable Long id, Principal principal) {
         String email = principal.getName();
         Optional<User> user = userRepository.findByEmail(email);
-
+        if (!user.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         Long userId = user.get().getId();
 
         WalletDTO walletDTO = walletService.findById(id, userId);
